s20e11

+++ resources/views/components/admin/ui/table-actions-list.blade.php
@@ -1,20 +1,28 @@
-@props(['item'])
+@props(['item', 'viewLink' => true, 'editLink' => true, 'deleteLink' => true])
 <td class="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
     <div class="flex items-center space-x-3.5 rtl:space-x-reverse">
-        <x-admin.ui.table-action-item href="{{ $item->viewLink() }}" target="_blank">
-            <x-admin.icon.eye/>
-        </x-admin.ui.table-action-item>
-        <x-admin.ui.table-action-item href="{{ $item->editLink() }}">
-            <x-admin.icon.edit/>
-        </x-admin.ui.table-action-item>
+        @if($viewLink)
+            <x-admin.ui.table-action-item href="{{ $item->viewLink() }}" target="_blank">
+                <x-admin.icon.eye/>
+            </x-admin.ui.table-action-item>
+        @endif
+        @if($editLink)
+            <x-admin.ui.table-action-item href="{{ $item->editLink() }}">
+                <x-admin.icon.edit/>
+            </x-admin.ui.table-action-item>
+        @endif
 
-        <form method="POST" action="{{ $item->deleteLink() }}">
-            @csrf
-            @method('DELETE')
-            <button type="submit" class="hover:text-primary">
-                <x-admin.icon.delete/>
-            </button>
-        </form>
+        {{ $slot }}
+
+        @if($deleteLink)
+            <form method="POST" action="{{ $item->deleteLink() }}">
+                @csrf
+                @method('DELETE')
+                <button type="submit" class="hover:text-primary">
+                    <x-admin.icon.delete/>
+                </button>
+            </form>
+        @endif
 
 
     </div>
